essentially remove all yaml include stuff
